+++ src/modules/menu/menu.service.ts
@@ -2,39 +2,73 @@ import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateMenuDto } from './dto/create-menu.dto';
 import { UpdateMenuDto } from './dto/update-menu.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
-import { Options } from 'src/types/OptionsFind';
-import { Prisma } from '@prisma/client';
+import { MenuOptionsPrisma } from 'src/prisma/helpers/options-prisma.type';
+import { TDayTime } from 'src/enums/daytime-menu';
 
 @Injectable()
 export class MenuService {
   constructor(private prisma: PrismaService) {}
 
-  create(body: CreateMenuDto) {
-    return this.prisma.menu.create({
-      data: body,
+  async create(body: CreateMenuDto) {
+    const menu = await this.prisma.menu.create({
+      data: {
+        name: body.name,
+        daytime: body.daytime,
+        MenuProduct: {
+          create: body.products.map((idProduct) => ({ idProduct: idProduct })),
+        },
+      },
+      include: {
+        MenuProduct: true,
+      },
     });
+    return menu;
   }
 
-  async findAll(options) {
-    const [menu, count] = await Promise.all([
-      this.prisma.menu.findMany(options),
-      this.prisma.menu.count({
-        where: options.where || {},
-      }),
-    ]);
-    return { rows: menu, count };
+  async daytime(daytime: TDayTime) {
+    const menu = await this.prisma.menu.findFirst({
+      where: {
+        daytime: daytime,
+      },
+      include: {
+        MenuProduct: {
+          select: {
+            product: {
+              select: {
+                name: true,
+                description: true,
+                price: true,
+              },
+            },
+          },
+        },
+      },
+    });
+
+    if (!menu) throw new NotFoundException('Menu não encontrado');
+
+    return menu;
   }
 
-  async findOne(optFind: {
-    where: object;
-    include?: Prisma.menuInclude;
-    select?: Prisma.menuSelect;
-  }) {
-    try {
-      return await this.prisma.menu.findFirstOrThrow(optFind);
-    } catch (error) {
-      throw new NotFoundException('Produto não encontrado');
-    }
+  async findAll() {
+    return this.prisma.menu.findMany();
+  }
+
+  async findOne(id: number) {
+    const menu = await this.findFirst({
+      where: { id },
+      include: {
+        MenuProduct: {
+          include: {
+            product: true,
+          },
+        },
+      },
+    });
+
+    if (!menu) throw new NotFoundException('Menu não encontrado');
+
+    return menu;
   }
 
   update(id: number, body: UpdateMenuDto) {
@@ -42,15 +76,33 @@ export class MenuService {
       where: {
         id: id,
       },
-      data: body,
+      data: {
+        name: body.name,
+        daytime: body.daytime,
+        // MenuProduct: {
+        //   connect: {
+        //     idMenu_idProduct:
+        //   },
+        // },
+      },
     });
   }
 
-  remove(id: number) {
+  async remove(id: number) {
+    const menuExists = await this.findFirst({
+      where: { id },
+    });
+
+    if (!menuExists) throw new NotFoundException('Menu não encontrado');
+
     return this.prisma.menu.delete({
       where: {
         id: id,
       },
     });
+  }
+
+  private findFirst(options: MenuOptionsPrisma) {
+    return this.prisma.menu.findFirst(options);
   }
 }
